final touchups

+++ export.py
@@ -3,11 +3,12 @@ import os
 
 
 def exportCSV(networklogs):
-    path = "static/download/output.csv"
+    path = "static/files/export.csv"
 
-    header = ['ip', 'hostname', 'country', 'lat', 'lng', 'org', 'attack']
+    header = ['srcip', 'sport', 'dstip', 'dport', 'hostname', 'country', 'lat', 'lng', 'org', 'attack']
 
-    os.remove(path)
+    if os.path.exists(path):
+        os.remove(path)
 
     with open(path, 'w', encoding='UTF8', newline='') as f:
         writer = csv.writer(f)
@@ -17,5 +18,5 @@ def exportCSV(networklogs):
 
         # write the data
         for networklog in networklogs:
-            data = [networklog.ip, networklog.hostname, networklog.country, networklog.lat, networklog.lng, networklog.org, networklog.attack]
+            data = [networklog.srcip, networklog.sport, networklog.dstip, networklog.dport, networklog.hostname, networklog.country, networklog.lat, networklog.lng, networklog.org, networklog.attack]
             writer.writerow(data)
